requirements.txt
fastapi
uvicorn
jinja2
sqlalchemy
aiofiles
python-multipart
passlib[bcrypt] 
openai
python-dotenv